mars/requirements.txt
black
mypy
dill
tqdm
pandas
requests
bs4
pillow
psycopg2-binary
torch
torchvision
